[add] test_polygamma

+++ cupy/special/zeta.py
@@ -109,12 +109,11 @@ def zeta(x, q):
             x.dtype == cupy.dtype('h') or x.dtype == cupy.dtype('B') or
             x.dtype == cupy.dtype('H') or x.dtype == cupy.bool_):
         x = x.astype(cupy.float32)
-        q = q.astype(cupy.float32)
     elif (x.dtype == cupy.dtype('i') or x.dtype == cupy.dtype('l') or
             x.dtype == cupy.dtype('q') or x.dtype == cupy.dtype('I') or
             x.dtype == cupy.dtype('L') or x.dtype == cupy.dtype('Q')):
         x = x.astype(cupy.float64)
-        q = q.astype(cupy.float64)
+    q = q.astype(x.dtype)
     x, q = cupy.asarray(x), cupy.asarray(q)
     x, q = cupy.broadcast_arrays(x, q)
     y = cupy.zeros_like(x)
